pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nsp-plugin"
version = "2.0.0"
description = "Neural Style Presets — AI-powered Lightroom preset prediction engine"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "torch>=2.2",
    "torchvision>=0.17",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "pandas>=2.0",
    "Pillow>=10.0",
    "opencv-python>=4.9",
    "rawpy>=0.23",
    "joblib>=1.4",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "SQLAlchemy>=2.0",
    "slowapi>=0.1",
    "timm>=0.9",
    "transformers>=4.40",
    "sentence-transformers>=3.0",
    "optuna>=3.6",
    "faiss-cpu>=1.8",
    "plotly>=5.0",
    "matplotlib>=3.8",
    "imagehash>=4.3",
    "PyWavelets>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "ruff>=0.1",
    "coverage>=7.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 60
addopts = "--tb=short"
